+++ CHANGES.md
@@ -1,6 +1,8 @@
 0.1.3
 
 - Using Spock 2.2 + Groovy 3 for tests
+- Add navigation link to Global Variables index page (only shown when global variables exist)
+
 0.1.2
 
 - Update maven-plugin dependencies to avoid vulnerability on plexus-utils@3.0.22